Move some hir_ty diagnostics to hir
